+++ examples/layout/sizing/lib/main.dart
@@ -1,0 +1,72 @@
+import 'package:flutter/material.dart';
+import 'package:flutter/rendering.dart' show debugPaintSizeEnabled;
+
+void main() {
+  debugPaintSizeEnabled = true; // Remove to suppress visual layout
+  runApp(MyApp());
+}
+
+class MyApp extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return MaterialApp(
+      title: 'Flutter layout demo',
+      home: Scaffold(
+        appBar: AppBar(
+          title: Text('Flutter layout demo'),
+        ),
+        // Change to buildFoo() for the other examples
+        body: Center(child: buildExpandedImages()),
+      ),
+    );
+  }
+
+  Widget buildOverflowingRow() =>
+      // #docregion overflowing-row
+      Row(
+        mainAxisAlignment: MainAxisAlignment.spaceEvenly,
+        children: [
+          Image.asset('images/pic1.jpg'),
+          Image.asset('images/pic2.jpg'),
+          Image.asset('images/pic3.jpg'),
+        ],
+      );
+  // #enddocregion overflowing-row
+
+  Widget buildExpandedImages() =>
+      // #docregion expanded-images
+      Row(
+        crossAxisAlignment: CrossAxisAlignment.center,
+        children: [
+          Expanded(
+            child: Image.asset('images/pic1.jpg'),
+          ),
+          Expanded(
+            child: Image.asset('images/pic2.jpg'),
+          ),
+          Expanded(
+            child: Image.asset('images/pic3.jpg'),
+          ),
+        ],
+      );
+  // #enddocregion expanded-images
+
+  Widget buildExpandedImagesWithFlex() =>
+      // #docregion expanded-images-with-flex
+      Row(
+        crossAxisAlignment: CrossAxisAlignment.center,
+        children: [
+          Expanded(
+            child: Image.asset('images/pic1.jpg'),
+          ),
+          Expanded(
+            flex: 2,
+            child: Image.asset('images/pic2.jpg'),
+          ),
+          Expanded(
+            child: Image.asset('images/pic3.jpg'),
+          ),
+        ],
+      );
+  // #enddocregion expanded-images-with-flex
+}
