--- src/_includes/code/layout/hello-world/pubspec.yaml
@@ -1,9 +1,0 @@
-name: hello_world
-description: Demonstrates a basic "Hello World" material app.
-
-dependencies:
-  flutter:
-    sdk: flutter
-
-flutter:
-  uses-material-design: true
