--- src/_includes/code/layout/row-expanded-2/pubspec.yaml
@@ -1,13 +1,0 @@
-name: row_expanded_2
-description: Demonstrates a Row widget containing an Expanded widget.
-
-dependencies:
-  flutter:
-    sdk: flutter
-
-flutter:
-  uses-material-design: true
-  assets:
-    - images/pic1.jpg
-    - images/pic2.jpg
-    - images/pic3.jpg
