--- src/_includes/code/layout/row-expanded/.gitignore
@@ -1,9 +1,0 @@
-.DS_Store
-.atom/
-.idea
-.packages
-.pub/
-build/
-ios/.generated/
-packages
-pubspec.lock
