--- src/_includes/code/layout/row-expanded/main.dart
@@ -1,62 +1,0 @@
-// Copyright 2017 The Chromium Authors. All rights reserved.
-// Use of this source code is governed by a BSD-style license that can be
-// found in the LICENSE file.
-
-import 'package:flutter/material.dart';
-// Comment out lines 7 and 10 to suppress the visual layout at runtime.
-import 'package:flutter/rendering.dart' show debugPaintSizeEnabled;
-
-void main() {
-  debugPaintSizeEnabled = true;
-  runApp(MyApp());
-}
-
-class MyApp extends StatelessWidget {
-  @override
-  Widget build(BuildContext context) {
-    return MaterialApp(
-      title: 'Flutter Demo',
-      theme: ThemeData(
-        primarySwatch: Colors.blue,
-      ),
-      home: MyHomePage(title: 'Flutter Demo Home Page'),
-    );
-  }
-}
-
-class MyHomePage extends StatefulWidget {
-  MyHomePage({Key key, this.title}) : super(key: key);
-
-  final String title;
-
-  @override
-  _MyHomePageState createState() => _MyHomePageState();
-}
-
-class _MyHomePageState extends State<MyHomePage> {
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: Text(widget.title),
-      ),
-      body: Center(
-        child: Row(
-          crossAxisAlignment: CrossAxisAlignment.center,
-          children: [
-            Expanded(
-              child: Image.asset('images/pic1.jpg'),
-            ),
-            Expanded(
-              flex: 2,
-              child: Image.asset('images/pic2.jpg'),
-            ),
-            Expanded(
-              child: Image.asset('images/pic3.jpg'),
-            ),
-          ],
-        ),
-      ),
-    );
-  }
-}
